# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: DZONE Shared Platform API
  version: 0.1.0
  description: Runtime-independent contract for workspaces, documents, and processing jobs.
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Service health.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /runtime/hosts/electron/open:
    post:
      operationId: launchElectronHost
      responses:
        "202":
          description: Electron host launch requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostLaunchResult"
  /runtime/hosts/electron/close:
    post:
      operationId: closeElectronHost
      responses:
        "202":
          description: Electron host close requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostLaunchResult"
  /runtime/hosts/dotnet-desktop/open:
    post:
      operationId: launchDotnetDesktopHost
      responses:
        "202":
          description: .NET Desktop host launch requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostLaunchResult"
  /runtime/hosts/dotnet-desktop/close:
    post:
      operationId: closeDotnetDesktopHost
      responses:
        "202":
          description: .NET Desktop host close requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostLaunchResult"
  /runtime/hosts/maui/open:
    post:
      operationId: launchMauiHost
      responses:
        "202":
          description: .NET MAUI desktop host launch requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostLaunchResult"
  /runtime/hosts/maui/close:
    post:
      operationId: closeMauiHost
      responses:
        "202":
          description: .NET MAUI desktop host close requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostLaunchResult"
  /runtime/backends/spring/open:
    post:
      operationId: launchSpringBackend
      responses:
        "202":
          description: Spring Boot backend launch requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostLaunchResult"
  /runtime/backends/spring/close:
    post:
      operationId: closeSpringBackend
      responses:
        "202":
          description: Spring Boot backend close requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostLaunchResult"
  /runtime/hosts/maui/setup:
    get:
      operationId: getMauiSetupStatus
      responses:
        "200":
          description: Optional .NET MAUI workload setup status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostSetupResult"
    post:
      operationId: setupMauiHost
      responses:
        "202":
          description: Optional .NET MAUI workload setup requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostSetupResult"
  /runtime/backends/spring/setup:
    get:
      operationId: getSpringSetupStatus
      responses:
        "200":
          description: Spring Boot prerequisite setup status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostSetupResult"
    post:
      operationId: setupSpringBackend
      responses:
        "202":
          description: Spring Boot prerequisite setup requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHostSetupResult"
  /workspaces:
    get:
      operationId: listWorkspaces
      responses:
        "200":
          description: Workspaces.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceInput"
      responses:
        "201":
          description: Created workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          $ref: "#/components/responses/NotFound"
  /documents:
    get:
      operationId: listDocuments
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Documents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentInput"
      responses:
        "201":
          description: Created document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateDocument
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentInput"
      responses:
        "200":
          description: Updated document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteDocument
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{id}/file:
    post:
      operationId: uploadDocumentFile
      parameters:
        - $ref: "#/components/parameters/Id"
        - name: x-file-name
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "202":
          description: File stored and processing job queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadDocumentFileResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{id}/process:
    post:
      operationId: processDocument
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  $ref: "#/components/schemas/ProcessingJobType"
      responses:
        "202":
          description: Processing job queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingJob"
        "404":
          $ref: "#/components/responses/NotFound"
  /jobs:
    get:
      operationId: listJobs
      parameters:
        - name: documentId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Processing jobs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcessingJob"
  /jobs/{id}:
    get:
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Processing job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingJob"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Health:
      type: object
      required: [ok, runtime]
      properties:
        ok:
          type: boolean
        runtime:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Workspace:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateWorkspaceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    DocumentStatus:
      type: string
      enum: [draft, review, approved, archived]
    Document:
      type: object
      required: [id, workspaceId, title, status, tags, createdAt, updatedAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        title:
          type: string
        status:
          $ref: "#/components/schemas/DocumentStatus"
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        fileName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
          minimum: 0
        fileStoredAt:
          type: string
          format: date-time
    CreateDocumentInput:
      type: object
      required: [workspaceId, title]
      properties:
        workspaceId:
          type: string
        title:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/DocumentStatus"
        tags:
          type: array
          items:
            type: string
    UpdateDocumentInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/DocumentStatus"
        tags:
          type: array
          items:
            type: string
    ProcessingJobType:
      type: string
      enum: [extract-text, thumbnail, summarize, index-search]
    ProcessingJobStatus:
      type: string
      enum: [queued, running, completed, failed]
    ProcessingJob:
      type: object
      required: [id, documentId, type, status, createdAt]
      properties:
        id:
          type: string
        documentId:
          type: string
        type:
          $ref: "#/components/schemas/ProcessingJobType"
        status:
          $ref: "#/components/schemas/ProcessingJobStatus"
        createdAt:
          type: string
          format: date-time
    UploadDocumentFileResult:
      type: object
      required: [document, job]
      properties:
        document:
          $ref: "#/components/schemas/Document"
        job:
          $ref: "#/components/schemas/ProcessingJob"
    RuntimeHostLaunchResult:
      type: object
      required: [host, status, backendUrl]
      properties:
        host:
          type: string
          enum: [electron, dotnet-desktop, maui-desktop, spring-backend]
        status:
          type: string
          enum: [starting, running, stopping, stopped]
        backendUrl:
          type: string
    RuntimeHostSetupResult:
      type: object
      required: [host, status, command]
      properties:
        host:
          type: string
          enum: [maui, spring]
        status:
          type: string
          enum: [idle, starting, running, completed, failed]
        command:
          type: string
        logPath:
          type: string
        startedAt:
          type: [string, "null"]
        finishedAt:
          type: [string, "null"]
        exitCode:
          type: [integer, "null"]
        signal:
          type: [string, "null"]
        lastOutput:
          type: string
        java:
          type: string
        maven:
          type: string
        spring:
          type: string